requirements.txt
flask
twilio
gspread
oauth2client
google-generativeai
python-dotenv